fix : 엔티티 관련 수정
- ManagerGroup BuilderDefault
- SaleRecord 요약 연결
- 전체매출 타입 변경

+++ common/src/main/java/com/kmhoon/common/model/entity/service/record/SaleRecordStatistics.java
@@ -26,7 +26,7 @@ public class SaleRecordStatistics {
     private LocalDate reportDate;
 
     @Comment("전체매출")
-    private Long totalRevenue;
+    private Double totalRevenue;
 
     @Comment("전월대비매출증감율")
     private Double mom;
